Bugfix in OctoAssociator.transform_events
In OctoAssociator.transform_events the events are transformed back from local to global EPSG:4326 coordinates. In that course a  scaling of the local coordinates from kilometers to meters or kilometers takes place, which requires a multiplication with the scaling factor (1000 or 1) rather than a division.

With this fix in place, correct event coordinates are now obtained also in EPSG:4326.